+++ packages/atproto/lib/src/atproto.dart
@@ -66,7 +66,7 @@ sealed class ATProto {
         ),
       );
 
-  /// Returns the merged headers with global headers and auth header.
+  /// Returns the global headers without auth header.
   Map<String, String> get headers;
 
   /// Returns the current session.